+++ src/main/java/cc/kostic/gec/instrument/OptionContract.java
@@ -2,10 +2,12 @@ package cc.kostic.gec.instrument;
 
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
+
 public class OptionContract extends Instrument {
 
 	private final String    option_type;	                // "call",
-	private final Double    strike;				            // 100000,
+	private final BigDecimal strike;				            // 100000,
 	
 	private Greeks greeks;
 
@@ -17,7 +19,7 @@ public class OptionContract extends Instrument {
 	public OptionContract(JSONObject o) {
 		super(o);
 		this.option_type   		 	        = o.getString("option_type");
-		this.strike    		 		        = o.getDouble("strike");
+		this.strike    		 		        = o.getBigDecimal("strike");
 	}
 
 
@@ -29,7 +31,7 @@ public class OptionContract extends Instrument {
 		}
 	}
 
-	public Double getStrike() {
+	public BigDecimal getStrike() {
 		return strike;
 	}
